fix: improve handling of data/lookup variables
